pc1_network_scanner/requirements.txt
python-nmap==0.7.1
pytest==8.0.0